requirements.txt
beautifulsoup4
requests
lxml
requests
playsound
python-dotenv
PyObjC
